+++ Src/Roflcopter.Plugin.Tests/Git/TestGitRepositoryProvider.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using Roflcopter.Plugin.Git;
+
+namespace Roflcopter.Plugin.Tests.Git
+{
+    internal class TestGitRepositoryProvider : GitRepositoryProvider
+    {
+        public TestGitRepositoryProvider(string gitDirectoryName = null)
+        {
+            GitDirectoryName = gitDirectoryName ?? $"__git_{Guid.NewGuid()}__";
+        }
+
+        public override string GitDirectoryName { get; }
+    }
+}