+++ Src/Roflcopter.Plugin/CopyFqnProviders/ShortNameTypeMemberFqnProvider.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using JetBrains.Annotations;
 using JetBrains.Application.DataContext;
+using JetBrains.Application.Settings;
 using JetBrains.ProjectModel;
 using JetBrains.ReSharper.Features.Environment.CopyFqn;
 using JetBrains.ReSharper.Psi;
@@ -16,29 +17,39 @@ namespace Roflcopter.Plugin.CopyFqnProviders
     [SolutionComponent]
     public class ShortNameTypeMemberFqnProvider : IFqnProvider
     {
+        private readonly ISolution _solution;
+
+        public ShortNameTypeMemberFqnProvider(ISolution solution)
+        {
+            _solution = solution;
+        }
+
         public bool IsApplicable([NotNull] IDataContext dataContext)
         {
             // Will be called in the CopyFqnAction to determine if _any_ provider has sth. to provide.
 
-            var typeMembers = GetTypeMembers(dataContext);
-
-            return typeMembers.Any();
+            return IsEnabled() && GetTypeMembers(dataContext).Any();
         }
 
         public IEnumerable<PresentableFqn> GetSortedFqns([NotNull] IDataContext dataContext)
         {
-            var typeMembers = GetTypeMembers(dataContext);
-
-            foreach (var typeMember in typeMembers)
+            if (IsEnabled())
             {
-                var typeMemberAndParents =
-                    GeneralUtil.SelfAndParentReversedPath<IClrDeclaredElement>(typeMember, x => x.GetContainingType()).Reverse();
+                var typeMembers = GetTypeMembers(dataContext);
 
-                yield return new PresentableFqn(string.Join(".", typeMemberAndParents.Select(x => x.ShortName)));
+                foreach (var typeMember in typeMembers)
+                {
+                    var typeMemberAndParents =
+                        GeneralUtil.SelfAndParentReversedPath<IClrDeclaredElement>(typeMember, x => x.GetContainingType()).Reverse();
+
+                    yield return new PresentableFqn(string.Join(".", typeMemberAndParents.Select(x => x.ShortName)));
+                }
             }
         }
 
         public int Priority => -10; // The lower the value, the _higher_ it is ranked. Yes, really.
+
+        private bool IsEnabled() => _solution.GetSettingsStore().GetValue((CopyFqnProvidersSettings s) => s.EnableShortNames);
 
         private static IEnumerable<ITypeMember> GetTypeMembers(IDataContext dataContext)
         {