+++ Src/Roflcopter.Plugin/Macros/GetBranchNameMacroImplementation.cs
@@ -1,28 +1,25 @@
-﻿using System.IO;
-using System.Linq;
+﻿using System.Linq;
 using System.Text.RegularExpressions;
 using JetBrains.Annotations;
 using JetBrains.ReSharper.Feature.Services.LiveTemplates.Hotspots;
 using JetBrains.ReSharper.Feature.Services.LiveTemplates.Macros;
-using JetBrains.Util;
+using JetBrains.ReSharper.Psi;
+using Roflcopter.Plugin.Git;
 
 namespace Roflcopter.Plugin.Macros
 {
     [MacroImplementation(Definition = typeof(GetBranchNameMacroDefinition))]
     public class GetBranchNameMacroImplementation : SimpleMacroImplementation
     {
-        private readonly IGetBranchNameMacroPathProvider _pathProvider;
-        private readonly string _gitDirectoryName;
+        private readonly GitRepositoryProvider _gitRepositoryProvider;
 
         [CanBeNull]
         private readonly IMacroParameterValueNew _stripingRegexParameter;
 
-        public GetBranchNameMacroImplementation(IGetBranchNameMacroPathProvider pathProvider, MacroParameterValueCollection parameters = null)
+        public GetBranchNameMacroImplementation(GitRepositoryProvider gitRepositoryProvider, MacroParameterValueCollection parameters = null)
         {
-            _pathProvider = pathProvider;
+            _gitRepositoryProvider = gitRepositoryProvider;
             _stripingRegexParameter = parameters.OptionalFirstOrDefault();
-
-            _gitDirectoryName = pathProvider.GitDirectoryName;
         }
 
         [CanBeNull]
@@ -33,53 +30,30 @@ namespace Roflcopter.Plugin.Macros
 
         private string Evaluate(IHotspotContext context)
         {
-            var sessionContextSolution = context.SessionContext.Solution;
-            var solutionDirectory = _pathProvider.GetSolutionDirectory(sessionContextSolution);
+            var psiSourceFile = context.SessionContext.Documents.First().GetPsiSourceFile(context.SessionContext.Solution);
+            var sourceFileLocation = psiSourceFile.GetLocation();
 
-            var findGitHeadFile = FindGitDirectory(solutionDirectory);
+            var gitRepositoryInfo = _gitRepositoryProvider.FetchGitRepositoryInfo(sourceFileLocation.Directory);
 
-            if (findGitHeadFile == null)
-                return $"<cannot find '{_gitDirectoryName}' directory>";
+            if (gitRepositoryInfo == null)
+                return "<cannot find Git repository directory>";
 
-            var headFile = findGitHeadFile.Combine("HEAD");
+            var headFileReference = gitRepositoryInfo.ReadHeadFileReference();
 
-            if (!headFile.ExistsFile)
-                return $"<cannot find '{_gitDirectoryName}/HEAD' file>";
+            if (headFileReference == null)
+                return "<cannot read reference in Git's HEAD file>";
 
-            var headFileContent = File.ReadAllLines(headFile.FullPath).DefaultIfEmpty("").First();
-
-            if (string.IsNullOrWhiteSpace(headFileContent))
-                return "<.git/HEAD file is empty>";
-
-            var headFileRef = Regex.Replace(headFileContent, "^ref: refs/heads/", "");
-
-            return ApplyStripingRegexArgument(headFileRef);
+            return ApplyStripingRegexArgument(headFileReference);
         }
 
-        private string ApplyStripingRegexArgument(string headFileRef)
+        private string ApplyStripingRegexArgument(string headFileReference)
         {
             var strippingRegex = _stripingRegexParameter != null ? _stripingRegexParameter.GetValue() : "";
 
             if (!string.IsNullOrWhiteSpace(strippingRegex))
-                return Regex.Replace(headFileRef, strippingRegex, "");
+                return Regex.Replace(headFileReference, strippingRegex, "");
 
-            return headFileRef;
-        }
-
-        [CanBeNull]
-        private FileSystemPath FindGitDirectory(FileSystemPath dir)
-        {
-            do
-            {
-                var gitDirectory = dir.Combine(_gitDirectoryName);
-
-                if (gitDirectory.ExistsDirectory)
-                    return gitDirectory;
-
-                dir = dir.Parent;
-            } while (!dir.IsEmpty);
-
-            return null;
+            return headFileReference;
         }
     }
 }